Add files via upload
The PBC3 clinical trial conducted in six European hospitals between 1983 and 1987 with primary biliary cirrhosis (PBC), where patients randomly assigned to receive Cyclosporin A (CyA) or a placebo. The main outcome became the "time to failure of medical treatment," and is defined as death or liver transplantation. This study dives deep into the factors that influence this time-to-failure, taking into account various patient characteristis recorded at the beginning of time.

Methods

The dataset 'Pbc3' was loaded with the “Import Dataset” method. I managed to give it a quick scan and evaluate any data that may be missing, and/or unique.
I loaded all of the necessary libraries: tidyr, dplyr, survival, ggplot2, and survminer. 
After loading the libraries, I performed data cleaning by checking for missing values and removing rows with missing data. This step makes sure that my analysis is based on complete records. I also calculated the number of patient records and the mean age of patients.
The dataset contains 275 patient records after removing rows with missing data with an average age of 54.15 years. The distribution of the status variable shows that 205 patients are censored, 24 are in the process of liver transplantation, and 46 patients died during the study.